tehtävä 4.21 tehty

+++ osa4/blogilista/controllers/blogs.js
@@ -60,8 +60,21 @@ blogsRouter.post('/', async (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+// hae pyynnöstä token ja sen käyttäjä kuten postissa
+// vertaa poistettavan blogin tekijän id
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  console.log('DecodedToken', decodedToken)   // tässä pyytäjän id
+  
+  const blog = await Blog.findById(request.params.id)
+  console.log('poistettava blogi', blog)
+
+  if(blog.user.toString() === decodedToken.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else
+  {
+    response.status(400).end()
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
